Add some it test. Refactor
