+++ conf/formsStack.conf
@@ -4,7 +4,7 @@ modules = [
   "formsKnot=io.knotx.forms.core.FormsKnot"
 
   ####### Forms Adpater #######
-  "formsAdapter=io.knotx.forms.adapter.mock.FormsMockAdapter"
+  "formsAdapter=io.knotx.forms.adapter.example.FormsExampleAdapter"
 ]
 
 ########### Globals ###########
@@ -12,7 +12,7 @@ global {
   address {
     forms {
       knot = knotx.knot.forms
-      mock.adapter = knotx.forms.mock.adapter.http
+      example.adapter = knotx.forms.example.adapter.http
     }
   }
 }
@@ -27,6 +27,6 @@ config.formsKnot {
 ########### Adapter configurations ###########
 config.formsAdapter {
   options.config {
-    address = ${global.address.forms.mock.adapter}
+    address = ${global.address.forms.example.adapter}
   }
 }
