+++ conf/includes/forms.conf
@@ -31,10 +31,10 @@ adapters = [
     # allowedResponseHeaders = []
   # }
 
-  ### Example mock forms adapter ###
+  ##### Example forms adapter #####
   {
     name = action-self
-    address = ${global.address.forms.mock.adapter}
+    address = ${global.address.forms.example.adapter}
     params = {
       example = example-value
     }
