+++ it-test/src/test/resources/formsStack_POST.conf
@@ -1,0 +1,9 @@
+include classpath("formsStack.conf")
+
+### Override Example Adapter Configuration ###
+config.formsAdapter {
+  options.config {
+    address = ${global.address.forms.example.adapter}
+    testStrategy = POST_REQUEST
+  }
+}
